🔨 chore: 카드 컴포넌트에 필요한 ProfileContainer 컴포넌트 제작
