--- src/components/card/profile/Profile.tsx
@@ -1,27 +1,0 @@
-import * as S from './Profile.style';
-
-type ProfileProps = {
-  name: string;
-  from: string;
-};
-
-const Profile: React.FC<ProfileProps> = ({ name, from }) => {
-  return (
-    <S.ProfileContainer>
-      <div
-        style={{
-          width: '40px',
-          height: '40px',
-          backgroundColor: 'var(--blue-200)',
-          borderRadius: '20px',
-        }}
-      ></div>
-      <S.ProfileTextContainer>
-        <S.NameItem>{name}</S.NameItem>
-        <S.InfoItem>{from}</S.InfoItem>
-      </S.ProfileTextContainer>
-    </S.ProfileContainer>
-  );
-};
-
-export default Profile;
